refactor: introduce @bitwarden/core-test-utils
