+++ .github/CODEOWNERS
@@ -99,6 +99,7 @@ libs/messaging-internal @bitwarden/team-platform-dev
 libs/serialization @bitwarden/team-platform-dev
 libs/guid @bitwarden/team-platform-dev
 libs/client-type @bitwarden/team-platform-dev
+libs/core-test-utils @bitwarden/team-platform-dev
 # Web utils used across app and connectors
 apps/web/src/utils/ @bitwarden/team-platform-dev
 # Web core and shared files
