+++ libs/core-test-utils/README.md
@@ -1,0 +1,5 @@
+# core-test-utils
+
+Owned by: platform
+
+Async test tools for state and clients
