+++ libs/core-test-utils/jest.config.js
@@ -1,0 +1,10 @@
+module.exports = {
+  displayName: "core-test-utils",
+  preset: "../../jest.preset.js",
+  testEnvironment: "node",
+  transform: {
+    "^.+\\.[tj]s$": ["ts-jest", { tsconfig: "<rootDir>/tsconfig.spec.json" }],
+  },
+  moduleFileExtensions: ["ts", "js", "html"],
+  coverageDirectory: "../../coverage/libs/core-test-utils",
+};
