+++ libs/core-test-utils/src/core-test-utils.spec.ts
@@ -1,0 +1,8 @@
+import * as lib from "./index";
+
+describe("core-test-utils", () => {
+  // This test will fail until something is exported from index.ts
+  it("should work", () => {
+    expect(lib).toBeDefined();
+  });
+});
