+++ tsconfig.base.json
@@ -28,6 +28,7 @@
       "@bitwarden/client-type": ["libs/client-type/src/index.ts"],
       "@bitwarden/common/*": ["./libs/common/src/*"],
       "@bitwarden/components": ["./libs/components/src"],
+      "@bitwarden/core-test-utils": ["libs/core-test-utils/src/index.ts"],
       "@bitwarden/dirt-card": ["./libs/dirt/card/src"],
       "@bitwarden/generator-components": ["./libs/tools/generator/components/src"],
       "@bitwarden/generator-core": ["./libs/tools/generator/core/src"],
